music-metadata-engine-backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
pydantic
supabase
sqlalchemy
psycopg2-binary
gotrue
email-validator
python-multipart
mutagen
numpy
librosa
google-generativeai
google-generativeai
